requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
boto3>=1.28.0
pyyaml>=6.0
fastapi>=0.103.0
uvicorn>=0.23.2
